move AepUI extension functions into a new ContentCardAepUIExtensions file
