+++ code/messaging/src/main/java/com/adobe/marketing/mobile/messaging/ContentCardMapper.kt
@@ -12,9 +12,6 @@
 package com.adobe.marketing.mobile.messaging
 
 import androidx.annotation.VisibleForTesting
-import com.adobe.marketing.mobile.MessagingEdgeEventType
-import com.adobe.marketing.mobile.aepcomposeui.AepUI
-import com.adobe.marketing.mobile.aepcomposeui.uimodels.SmallImageTemplate
 import com.adobe.marketing.mobile.util.StringUtils
 
 /**
@@ -68,33 +65,5 @@ class ContentCardMapper private constructor() {
     @VisibleForTesting
     internal fun clear() {
         contentCardSchemaDataMap.clear()
-    }
-}
-
-/**
- * Extension function to get the meta data for the given [AepUI].
- *
- * @return the meta data as a [MutableMap] or null if the [AepUI] does not have meta data.
- */
-
-fun AepUI<*, *>.getMeta(): Map<String, Any>? {
-    return when (val template = this.getTemplate()) {
-        is SmallImageTemplate ->
-            ContentCardMapper.instance.getContentCardSchemaData(template.id)?.meta
-
-        else -> null
-    }
-}
-
-/**
- * Extension function to track a custom interaction for the given [AepUI].
- *
- * @param interaction the custom interaction to track
- */
-fun AepUI<*, *>.trackInteraction(interaction: String) {
-    when (val template = this.getTemplate()) {
-        is SmallImageTemplate ->
-            ContentCardMapper.instance.getContentCardSchemaData(template.id)
-                ?.track(interaction, MessagingEdgeEventType.INTERACT)
     }
 }
